requirements.txt
librosa
transformers
python-dotenv
sounddevice
scipy
keyboard